Remove homepage and use dashboard as default
