+++ src/App.tsx
@@ -8,7 +8,6 @@ import LoginPage from './pages/LoginPage'
 import Dashboard from './pages/Dashboard'
 import Schedule from './pages/Schedule'
 import Workforce from './pages/Workforce'
-import Homepage from './pages/Homepage'
 import ProtectedRoute from './auth/ProtectedRoute'
 import MainLayout from './layout/MainLayout'
 import AdmissionLayout from './pages/admission/AdmissionLayout'
@@ -22,8 +21,7 @@ function App() {
         <Route path="/login" element={<LoginPage />} />
         <Route element={<ProtectedRoute />}>
           <Route path="/" element={<MainLayout />}>
-            <Route index element={<Navigate to="homepage" replace />} />
-            <Route path="homepage" element={<Homepage />} />
+            <Route index element={<Navigate to="dashboard" replace />} />
             <Route path="dashboard" element={<Dashboard />} />
             <Route path="schedule" element={<Schedule />} />
             <Route path="workforce" element={<Workforce />} />
