+++ src/pages/Dashboard.tsx
@@ -1,7 +1,22 @@
+import React from 'react'
+import { useNavigate } from 'react-router-dom'
+import { useAuth } from '../auth/AuthContext'
+
 const Dashboard: React.FC = () => {
+  const { user, logout } = useAuth()
+  const navigate = useNavigate()
+
+  const handleLogout = () => {
+    logout()
+    navigate('/login')
+  }
+
   return (
-    <div>
-      <h1>Dashboard Page</h1>
+    <div className="dashboard">
+      <div className="content">
+        <h1>Welcome, {user?.name}</h1>
+        <p>This is your admin dashboard.</p>
+      </div>
     </div>
   )
 }
